src_analytics/pyproject.toml
[project]
name         = "analytics"
description  = "Internet of Things Predictive Modelling Libraries"
authors      = [{name = "Joshoua Bigler"}]
version      = "0.0.1"
dependencies = [
  "numpy>=2.2.0",
  "mlflow>=2.21.3",
  "pandas>=2.2.3",
  "matplotlib>=3.10.1",
  "seaborn>=0.13.2",
  "scikit-learn>=1.6.1",
  "shap>=0.47.1",
  "ipykernel>=6.29.5",
  "ipywidgets>=8.1.5",
  "wandb>=0.19.8",
  "torchinfo>=1.8.0",
  "nbformat>=5.10.4"
]

[tool.setuptools]
package-dir  = { "" = "." }
packages = { find = { include = ["analytics*"] } } 
[tool.pytest.ini_options]
testpaths = ["test"]